+++ tutorials/fannie_mae_new/01d_OPTIONAL_rechunk.r
@@ -1,5 +1,0 @@
-library(disk.frame)
-
-fmdf <- disk.frame("fmdf")
-
-rechunk(fmdf, 128)